apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
httpx==0.25.2
web3==6.11.3
python-dotenv==1.0.0
